+++ resources/js/components/Clients.vue
@@ -224,6 +224,35 @@
                 </div>
             </div>
         </div>
+
+        <!-- Client Secret Modal -->
+        <div class="modal fade" id="modal-client-secret" tabindex="-1" role="dialog">
+            <div class="modal-dialog">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h4 class="modal-title">
+                            Client Secret
+                        </h4>
+
+                        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+                    </div>
+
+                    <div class="modal-body">
+                        <p>
+                            Here is your new client secret. This is the only time it will be shown so don't lose it!
+                            You may now use this secret to make API requests.
+                        </p>
+
+                        <input type="text" class="form-control" v-model="clientSecret">
+                    </div>
+
+                    <!-- Modal Actions -->
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                    </div>
+                </div>
+            </div>
+        </div>
     </div>
 </template>
 
@@ -235,6 +264,8 @@
         data() {
             return {
                 clients: [],
+
+                clientSecret: null,
 
                 createForm: {
                     errors: [],
@@ -303,8 +334,10 @@
              */
             store() {
                 this.persistClient(
-                    'post', '/oauth/clients',
-                    this.createForm, '#modal-create-client'
+                    'post',
+                    '/oauth/clients',
+                    this.createForm,
+                    '#modal-create-client'
                 );
             },
 
@@ -324,8 +357,10 @@
              */
             update() {
                 this.persistClient(
-                    'put', '/oauth/clients/' + this.editForm.id,
-                    this.editForm, '#modal-edit-client'
+                    'put',
+                    '/oauth/clients/' + this.editForm.id,
+                    this.editForm,
+                    '#modal-edit-client'
                 );
             },
 
@@ -337,13 +372,17 @@
 
                 axios[method](uri, form)
                     .then(response => {
-                        this.clients.push(response.data);
+                        this.getClients();
 
                         form.name = '';
                         form.redirect = '';
                         form.errors = [];
 
                         $(modal).modal('hide');
+
+                        if (response.data.plainSecret) {
+                            this.showClientSecret(response.data.plainSecret);
+                        }
                     })
                     .catch(error => {
                         if (typeof error.response.data === 'object') {
@@ -352,6 +391,15 @@
                             form.errors = ['Something went wrong. Please try again.'];
                         }
                     });
+            },
+
+            /**
+             * Show the given client secret to the user.
+             */
+            showClientSecret(clientSecret) {
+                this.clientSecret = clientSecret;
+
+                $('#modal-client-secret').modal('show');
             },
 
             /**