+++ src/Console/ClientCommand.php
@@ -5,6 +5,7 @@ namespace Laravel\Passport\Console;
 use Illuminate\Console\Command;
 use Laravel\Passport\Client;
 use Laravel\Passport\ClientRepository;
+use Laravel\Passport\Passport;
 
 class ClientCommand extends Command
 {
@@ -161,6 +162,11 @@ class ClientCommand extends Command
      */
     protected function outputClientDetails(Client $client)
     {
+        if (Passport::$hashesClientSecrets) {
+            $this->line('<comment>Here is your new client secret. This is the only time it will be shown so don\'t lose it!</comment>');
+            $this->line('');
+        }
+
         $this->line('<comment>Client ID:</comment> '.$client->id);
         $this->line('<comment>Client secret:</comment> '.$client->plainSecret);
     }