+++ src/Console/HashCommand.php
@@ -12,7 +12,7 @@ class HashCommand extends Command
      *
      * @var string
      */
-    protected $signature = 'passport:hash';
+    protected $signature = 'passport:hash {--force : Force the operation to run without confirmation prompt}';
 
     /**
      * The console command description.
@@ -34,7 +34,7 @@ class HashCommand extends Command
             return;
         }
 
-        if ($this->confirm('Are you sure you want to hash all client secrets? This cannot be undone.')) {
+        if ($this->option('force') || $this->confirm('Are you sure you want to hash all client secrets? This cannot be undone.')) {
             $model = Passport::clientModel();
 
             foreach ((new $model)->whereNotNull('secret')->cursor() as $client) {